requirements.txt
pip == 21.1.1
PyJWT == 2.9.0
Django == 4.2.16
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.3.1
pytest == 8.3.3
pytest-django == 4.9.0
tzdata == 2024.2
sqlparse == 0.5.1
pymemcache == 4.0.0
pillow == 10.4.0